Pass language name instead of locale name to `get_javascript_catalog`.
Even though the argument is named locale, it makes its way into
`DjangoTranslation`, which expects it to be `language`, ultimately calls
`to_locale` on it. This is not a big issue for unhyphenated languages.

However, it fails silently for languages like zh-hans, resulting in an
empty catalog. This commit addresses the issue, and adds a `zh-hans`
test case to catch this failure.